move store decoders to modules
