+++ x/auth/simulation/store.go
@@ -1,0 +1,31 @@
+package simulation
+
+import (
+	"bytes"
+	"fmt"
+
+	cmn "github.com/tendermint/tendermint/libs/common"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	"github.com/cosmos/cosmos-sdk/x/auth"
+	"github.com/cosmos/cosmos-sdk/x/auth/exported"
+)
+
+// DecodeStore unmarshals the KVPair's Value to the corresponding auth type
+func DecodeStore(cdcA, cdcB *codec.Codec, kvA, kvB cmn.KVPair) string {
+	switch {
+	case bytes.Equal(kvA.Key[:1], auth.AddressStoreKeyPrefix):
+		var accA, accB exported.Account
+		cdcA.MustUnmarshalBinaryBare(kvA.Value, &accA)
+		cdcB.MustUnmarshalBinaryBare(kvB.Value, &accB)
+		return fmt.Sprintf("%v\n%v", accA, accB)
+	case bytes.Equal(kvA.Key, auth.GlobalAccountNumberKey):
+		var globalAccNumberA, globalAccNumberB uint64
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &globalAccNumberA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &globalAccNumberB)
+		return fmt.Sprintf("GlobalAccNumberA: %d\nGlobalAccNumberB: %d", globalAccNumberA, globalAccNumberB)
+	default:
+		panic(fmt.Sprintf("invalid account key %X", kvA.Key))
+	}
+}
