+++ x/distribution/simulation/store.go
@@ -1,0 +1,69 @@
+package simulation
+
+import (
+	"bytes"
+	"fmt"
+
+	cmn "github.com/tendermint/tendermint/libs/common"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/distribution"
+)
+
+// DecodeStore unmarshals the KVPair's Value to the corresponding distribution type
+func DecodeStore(cdcA, cdcB *codec.Codec, kvA, kvB cmn.KVPair) string {
+	switch {
+	case bytes.Equal(kvA.Key[:1], distribution.FeePoolKey):
+		var feePoolA, feePoolB distribution.FeePool
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &feePoolA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &feePoolB)
+		return fmt.Sprintf("%v\n%v", feePoolA, feePoolB)
+
+	case bytes.Equal(kvA.Key[:1], distribution.ProposerKey):
+		return fmt.Sprintf("%v\n%v", sdk.ConsAddress(kvA.Value), sdk.ConsAddress(kvB.Value))
+
+	case bytes.Equal(kvA.Key[:1], distribution.ValidatorOutstandingRewardsPrefix):
+		var rewardsA, rewardsB distribution.ValidatorOutstandingRewards
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &rewardsA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &rewardsB)
+		return fmt.Sprintf("%v\n%v", rewardsA, rewardsB)
+
+	case bytes.Equal(kvA.Key[:1], distribution.DelegatorWithdrawAddrPrefix):
+		return fmt.Sprintf("%v\n%v", sdk.AccAddress(kvA.Value), sdk.AccAddress(kvB.Value))
+
+	case bytes.Equal(kvA.Key[:1], distribution.DelegatorStartingInfoPrefix):
+		var infoA, infoB distribution.DelegatorStartingInfo
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &infoA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &infoB)
+		return fmt.Sprintf("%v\n%v", infoA, infoB)
+
+	case bytes.Equal(kvA.Key[:1], distribution.ValidatorHistoricalRewardsPrefix):
+		var rewardsA, rewardsB distribution.ValidatorHistoricalRewards
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &rewardsA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &rewardsB)
+		return fmt.Sprintf("%v\n%v", rewardsA, rewardsB)
+
+	case bytes.Equal(kvA.Key[:1], distribution.ValidatorCurrentRewardsPrefix):
+		var rewardsA, rewardsB distribution.ValidatorCurrentRewards
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &rewardsA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &rewardsB)
+		return fmt.Sprintf("%v\n%v", rewardsA, rewardsB)
+
+	case bytes.Equal(kvA.Key[:1], distribution.ValidatorAccumulatedCommissionPrefix):
+		var commissionA, commissionB distribution.ValidatorAccumulatedCommission
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &commissionA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &commissionB)
+		return fmt.Sprintf("%v\n%v", commissionA, commissionB)
+
+	case bytes.Equal(kvA.Key[:1], distribution.ValidatorSlashEventPrefix):
+		var eventA, eventB distribution.ValidatorSlashEvent
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &eventA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &eventB)
+		return fmt.Sprintf("%v\n%v", eventA, eventB)
+
+	default:
+		panic(fmt.Sprintf("invalid distribution key prefix %X", kvA.Key[:1]))
+	}
+}
