+++ x/mint/simulation/store.go
@@ -1,0 +1,25 @@
+package simulation
+
+import (
+	"bytes"
+	"fmt"
+
+	cmn "github.com/tendermint/tendermint/libs/common"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	"github.com/cosmos/cosmos-sdk/x/mint"
+)
+
+// DecodeStore unmarshals the KVPair's Value to the corresponding mint type
+func DecodeStore(cdcA, cdcB *codec.Codec, kvA, kvB cmn.KVPair) string {
+	switch {
+	case bytes.Equal(kvA.Key, mint.MinterKey):
+		var minterA, minterB mint.Minter
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &minterA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &minterB)
+		return fmt.Sprintf("%v\n%v", minterA, minterB)
+	default:
+		panic(fmt.Sprintf("invalid mint key %X", kvA.Key))
+	}
+}
