+++ x/mint/simulation/store_test.go
@@ -1,0 +1,50 @@
+package simulation
+
+import (
+	"fmt"
+	"testing"
+
+	"github.com/stretchr/testify/require"
+
+	cmn "github.com/tendermint/tendermint/libs/common"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	"github.com/cosmos/cosmos-sdk/x/mint"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
+
+func makeTestCodec() (cdc *codec.Codec) {
+	cdc = codec.New()
+	sdk.RegisterCodec(cdc)
+	return
+}
+
+func TestDecodeStore(t *testing.T) {
+	cdc := makeTestCodec()
+	minter := mint.NewMinter(sdk.OneDec(), sdk.NewDec(15))
+
+	kvPairs := cmn.KVPairs{
+		cmn.KVPair{Key: mint.MinterKey, Value: cdc.MustMarshalBinaryLengthPrefixed(minter)},
+		cmn.KVPair{Key: []byte{0x99}, Value: []byte{0x99}},
+	}
+	tests := []struct {
+		name        string
+		expectedLog string
+	}{
+		{"Minter", fmt.Sprintf("%v\n%v", minter, minter)},
+		{"other", ""},
+	}
+
+	for i, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			switch i {
+			case len(tests) - 1:
+				require.Panics(t, func() { DecodeStore(cdc, cdc, kvPairs[i], kvPairs[i]) }, tt.name)
+			default:
+				require.Equal(t, tt.expectedLog, DecodeStore(cdc, cdc, kvPairs[i], kvPairs[i]), tt.name)
+			}
+		})
+	}
+}
