+++ x/slashing/simulation/store.go
@@ -1,0 +1,42 @@
+package simulation
+
+import (
+	"bytes"
+	"fmt"
+
+	"github.com/tendermint/tendermint/crypto"
+	cmn "github.com/tendermint/tendermint/libs/common"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/slashing"
+)
+
+// DecodeStore unmarshals the KVPair's Value to the corresponding slashing type
+func DecodeStore(cdcA, cdcB *codec.Codec, kvA, kvB cmn.KVPair) string {
+	switch {
+	case bytes.Equal(kvA.Key[:1], slashing.ValidatorSigningInfoKey):
+		var infoA, infoB slashing.ValidatorSigningInfo
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &infoA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &infoB)
+		return fmt.Sprintf("%v\n%v", infoA, infoB)
+
+	case bytes.Equal(kvA.Key[:1], slashing.ValidatorMissedBlockBitArrayKey):
+		var missedA, missedB bool
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &missedA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &missedB)
+		return fmt.Sprintf("missedA: %v\nmissedB: %v", missedA, missedB)
+
+	case bytes.Equal(kvA.Key[:1], slashing.AddrPubkeyRelationKey):
+		var pubKeyA, pubKeyB crypto.PubKey
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &pubKeyA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &pubKeyB)
+		bechPKA := sdk.MustBech32ifyAccPub(pubKeyA)
+		bechPKB := sdk.MustBech32ifyAccPub(pubKeyB)
+		return fmt.Sprintf("PubKeyA: %s\nPubKeyB: %s", bechPKA, bechPKB)
+
+	default:
+		panic(fmt.Sprintf("invalid slashing key prefix %X", kvA.Key[:1]))
+	}
+}
