+++ x/staking/simulation/store.go
@@ -1,0 +1,58 @@
+package simulation
+
+import (
+	"bytes"
+	"fmt"
+
+	cmn "github.com/tendermint/tendermint/libs/common"
+
+	"github.com/cosmos/cosmos-sdk/codec"
+
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/x/staking"
+)
+
+// DecodeStore unmarshals the KVPair's Value to the corresponding staking type
+func DecodeStore(cdcA, cdcB *codec.Codec, kvA, kvB cmn.KVPair) string {
+	switch {
+	case bytes.Equal(kvA.Key[:1], staking.LastTotalPowerKey):
+		var powerA, powerB sdk.Int
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &powerA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &powerB)
+		return fmt.Sprintf("%v\n%v", powerA, powerB)
+
+	case bytes.Equal(kvA.Key[:1], staking.ValidatorsKey):
+		var validatorA, validatorB staking.Validator
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &validatorA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &validatorB)
+		return fmt.Sprintf("%v\n%v", validatorA, validatorB)
+
+	case bytes.Equal(kvA.Key[:1], staking.LastValidatorPowerKey),
+		bytes.Equal(kvA.Key[:1], staking.ValidatorsByConsAddrKey),
+		bytes.Equal(kvA.Key[:1], staking.ValidatorsByPowerIndexKey):
+		return fmt.Sprintf("%v\n%v", sdk.ValAddress(kvA.Value), sdk.ValAddress(kvB.Value))
+
+	case bytes.Equal(kvA.Key[:1], staking.DelegationKey):
+		var delegationA, delegationB staking.Delegation
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &delegationA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &delegationB)
+		return fmt.Sprintf("%v\n%v", delegationA, delegationB)
+
+	case bytes.Equal(kvA.Key[:1], staking.UnbondingDelegationKey),
+		bytes.Equal(kvA.Key[:1], staking.UnbondingDelegationByValIndexKey):
+		var ubdA, ubdB staking.UnbondingDelegation
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &ubdA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &ubdB)
+		return fmt.Sprintf("%v\n%v", ubdA, ubdB)
+
+	case bytes.Equal(kvA.Key[:1], staking.RedelegationKey),
+		bytes.Equal(kvA.Key[:1], staking.RedelegationByValSrcIndexKey):
+		var redA, redB staking.Redelegation
+		cdcA.MustUnmarshalBinaryLengthPrefixed(kvA.Value, &redA)
+		cdcB.MustUnmarshalBinaryLengthPrefixed(kvB.Value, &redB)
+		return fmt.Sprintf("%v\n%v", redA, redB)
+
+	default:
+		panic(fmt.Sprintf("invalid staking key prefix %X", kvA.Key[:1]))
+	}
+}
